pyproject.toml
[tool.poetry]
name = "marques-bot"
version = "0.1.5"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.5"
python-dotenv = "^1.0.1"
yt-dlp = "^2024.8.6"
telethon = "^1.36.0"
setuptools = "^74.1.2"
cryptg = "^0.4.0"
pillow = "^10.4.0"
hachoir = "^3.3.0"
urwid = "^2.6.15"
ffmpy = "^0.4.0"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
